utils/requirements.txt
tqdm
scipy
numpy
GitPython
pybids
nilearn
sklearn
matplotlib
addict
nibabel
tabulate
comet-ml
pypmi
pytorch-ignite==0.4.4
pyvista==0.30.1
itkwidgets
openpyxl
monai
pytorch3dunet @ git+https://github.com/wolny/pytorch-3dunet.git
torch
torchvision
torchaudio
torch-sparse
torch-cluster
torch-scatter
torch-geometric==1.7.0
